requirements.txt
aws-cdk.core>=1.29.0
aws-cdk-aws.iam>=1.29.0
aws-cdk.aws-s3>=1.29.0
aws-cdk.aws-s3.deployment>=1.29.0
aws-cdk.aws-kms>=1.29.0
aws-cdk.aws-ec2>=1.29.0
aws-cdk.aws-emr>=1.29.0
aws-cdk.aws-sns>=1.29.0
aws-cdk.aws-sqs>=1.29.0
aws-cdk.aws-ssm>=1.29.0
aws-cdk.aws-secretsmanager>=1.29.0
aws-cdk.aws-lambda>=1.29.0
aws-cdk.aws-lambda-event-sources>=1.29.0
aws-cdk.aws-stepfunctions>=1.29.0
aws-cdk.aws-stepfunctions-tasks>=1.29.0
aws-cdk.aws-events>=1.29.0
aws-cdk.aws-events-targets>=1.29.0
boto3>=1.12.23
logzero~=1.5.0
